requirements-server.txt
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
python-engineio>=4.8.0
Pillow>=10.0.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
psutil>=5.9.0
python-dotenv>=1.0.0
eventlet>=0.33.0
